Gave admin permission to edit Question objects
Gave admin permission to edit Question objects. 
PS note sure were to state this, but before doing so I created a superuser and created choices. For some reason I cannot find what files this information is stored in. Assume that I am doing the command lines as instructed in tutorial when not stated otherwise.