command line update
This is the result of the stuff done on the command line as well as the stuff done on the website. It basically is where all of the data is stored. I am only uploading this now because I did not realize this is where that was being stored. Currently, I have made 1 question, gave it 2 choices, created an admin, and changed the date on the question.