Changes made to find bug
Found the bug (that future times are counted as being published recently). to find this required future_question variable to be made and had a value stored into it. This is the only change that I know of that would have been made from what I did on the command line.